+++ src/main/java/tf/gpx/edit/main/GPXEditor.java
@@ -121,7 +121,6 @@ import tf.gpx.edit.algorithms.WaypointClustering;
 import tf.gpx.edit.algorithms.WaypointReduction;
 import tf.gpx.edit.elevation.AssignElevation;
 import tf.gpx.edit.elevation.FindElevation;
-import tf.gpx.edit.panorama.PanoramaViewer;
 import tf.gpx.edit.elevation.SRTMDataViewer;
 import tf.gpx.edit.elevation.SRTMDownloader;
 import tf.gpx.edit.extension.LineStyle;
@@ -148,6 +147,7 @@ import tf.gpx.edit.items.GPXWaypoint;
 import tf.gpx.edit.leafletmap.IGeoCoordinate;
 import tf.gpx.edit.leafletmap.LatLonElev;
 import tf.gpx.edit.leafletmap.MapLayerUsage;
+import tf.gpx.edit.panorama.PanoramaViewer;
 import tf.gpx.edit.values.DistributionViewer;
 import tf.gpx.edit.values.EditGPXMetadata;
 import tf.gpx.edit.values.EditGPXWaypoint;
