+++ src/main/java/tf/gpx/edit/sun/SunPathForDate.java
@@ -29,7 +29,6 @@ import java.time.Duration;
 import java.time.format.DateTimeFormatter;
 import java.util.Calendar;
 import java.util.GregorianCalendar;
-import java.util.TimeZone;
 import org.apache.commons.lang3.time.DurationFormatUtils;
 import tf.helper.general.ObjectsHelper;
