+++ src/main/java/tf/gpx/edit/viewer/IChartBasics.java
@@ -33,7 +33,6 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.stream.Collectors;
 import javafx.application.Platform;
-import javafx.css.PseudoClass;
 import javafx.geometry.BoundingBox;
 import javafx.geometry.Insets;
 import javafx.scene.CacheHint;
